Removes some dead streamquery code, adds Apache license doc blocks, restructures some stuff
