+++ streamalert/streamquery/command/processor.py
@@ -1,3 +1,18 @@
+"""
+Copyright 2017-present, Airbnb Inc.
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+   http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+"""
 from logging import Logger
 
 from streamalert.streamquery.query_packs.manager import QueryPacksManager
@@ -67,7 +82,6 @@ class CommandProcessor:
             )
             self._logger.error(query_execution.status_description)
 
-            # FIXME (derek.wang) No rules are written yet https://jira.airbnb.biz/browse/CSIRT-1031
             self._kinesis.send_error_results(query_pack)
 
             self._mark_query_pack_sent(query_pack)
